Hiding log behind debug feature
